fix: getQuote separated types
